temp/requirements.txt
Django==4.2
channels==4.0
daphne==4.0
openai
torch
torchaudio
openai-whisper
